std::string --> UiRadioButtons::append
