+++ src/UiRadioButtons.cc
@@ -1,6 +1,19 @@
+#include <string>
 #include "../ui.h"
 #include "ui-node.h"
 #include "nbind/api.h"
+
+class UiRadioButtons : public UiControl {
+	DEFINE_EVENT(onSelected)
+
+  public:
+	UiRadioButtons();
+	void append(std::string text);
+	int getSelected();
+	void setSelected(int n);
+
+	DEFINE_CONTROL_METHODS()
+};
 
 UiRadioButtons::UiRadioButtons()
 	: UiControl((uiControl *)uiNewRadioButtons()) {}
@@ -21,8 +34,8 @@ void UiRadioButtons::setSelected(int n) {
 	}
 }
 
-void UiRadioButtons::append(const char *text) {
-	uiRadioButtonsAppend((uiRadioButtons *)getHandle(), text);
+void UiRadioButtons::append(std::string text) {
+	uiRadioButtonsAppend(uiRadioButtons(getHandle()), text.c_str());
 }
 
 NBIND_CLASS(UiRadioButtons) {
